% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplitude.R
\name{amplitude}
\alias{amplitude}
\title{Compute amplitude from Fourier parameters}
\usage{
amplitude(x, coordinate = c("x", "y", "all"))
}
\arguments{
\item{x}{A data frame or matrix with columns \code{an}, \code{bn}, \code{cn}, and \code{dn}.}

\item{coordinate}{Character. Specifies whether to compute amplitude for the "x" coordinate, "y" coordinate, or "all". Must be one of \code{"x"}, \code{"y"}, or \code{"all"}.}
}
\value{
A numeric vector containing amplitude values for each harmonic.
}
\description{
Computes the amplitude from a given set of Fourier parameters \code{an}, \code{bn}, \code{cn}, and \code{dn}.
}
