% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpol_s.R
\name{interpol_s}
\alias{interpol_s}
\title{Interpolate shape given two sets of Fourier parameters}
\usage{
interpol_s(
  shape1,
  shape2,
  method = c("mean", "sequence"),
  n.shapes = NULL,
  npts,
  fou.pars = F
)
}
\arguments{
\item{shape1}{A data frame or matrix with columns \code{an}, \code{bn}, \code{cn}, and \code{dn}. It can also be a list with elements \code{an}, \code{bn}, \code{cn} and \code{dn}}

\item{shape2}{A data frame or matrix with columns \code{an}, \code{bn}, \code{cn}, and \code{dn}. It can also be a list with elements \code{an}, \code{bn}, \code{cn} and \code{dn}}

\item{method}{Character. Specifies whether a single mean shape is computed or a sequence of intermediate shapes between shape1 and shape2. More methods will be implemented in the future}

\item{npts}{Integer. Number of points to use in the reconstructed shape.}

\item{fou.pars}{Logical. If \code{TRUE}, the function returns a list containing the Fourier parameters and the reconstructed shape coordinates. Default is \code{FALSE}.}

\item{n.shapes.}{If method = "sequence", the number of interpolated shapes}
}
\value{
If \code{fou.pars = FALSE}, returns a matrix of shape coordinates (x,y).
If \code{fou.pars = TRUE}, returns a list with two elements:
\enumerate{
\item \code{Fourier_parameters}: data frame of Fourier coefficients used, with rows for harmonics and columns \code{an}, \code{bn}, \code{cn}, and \code{dn}.
\item \code{Shape}: matrix of shape coordinates (x,y).
If \code{method = "sequence"}, returns a list with length = sequence. Each element within the list is structured as above
}
}
\description{
Computes an interpolated shape from a given set of Fourier parameters \code{an}, \code{bn}, \code{cn}, and \code{dn}.
}
