% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase.R
\name{phase}
\alias{phase}
\title{Compute phase from Fourier parameters}
\usage{
phase(x, coordinate = c("x", "y"))
}
\arguments{
\item{x}{A data frame or matrix with columns \code{an}, \code{bn}, \code{cn}, and \code{dn}.}

\item{coordinate}{Character. Specifies whether to compute phase for the \code{"x"} or \code{"y"} coordinate. Must be one of \code{"x"} or \code{"y"}.}
}
\value{
A numeric vector containing phase values for each harmonic.
}
\description{
Computes the phase from a given set of Fourier parameters \code{an}, \code{bn}, \code{cn}, and \code{dn}.
}
