% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simumorph.R
\name{simumorph}
\alias{simumorph}
\title{Simulate shape evolution in morphospace given a covariance matrix}
\usage{
simumorph(
  x,
  m.space,
  init,
  init.from.morphospace = T,
  target,
  target.from.morphospace = T,
  method = c("AtoA", "AtoB", "AtoMult", "Free"),
  sim,
  npts,
  a = 0.2,
  e = 0.15,
  dynamic_e = F,
  f = 100,
  int.allowed = F,
  only.shapes = F,
  max.attempts = 500
)
}
\arguments{
\item{x}{A positive definite covariance matrix of shape parameters from the morphospace.}

\item{m.space}{A data frame of amplitude and phase values for all shapes in the morphospace (ideally from \code{morphospace()}).}

\item{init}{Integer or character. The initial shape's row index or name in \code{m.space} where simulation begins.}

\item{init.from.morphospace}{Logical. Whether to include the initial shape in truncations. Default is \code{TRUE}.}

\item{target}{Integer, character, or vector thereof. Row index(es) or name(s) in \code{m.space} of target shape(s) guiding the simulation.}

\item{target.from.morphospace}{Logical. Whether to include target shape(s) in truncations. Default is \code{TRUE}.}

\item{method}{Character. Simulation mode; one of \code{"AtoA"}, \code{"AtoB"}, \code{"AtoMult"}, or \code{"Free"}. Defaults to \code{"AtoA"}.
\describe{
\item{\code{"AtoA"}}{Simulation stays near the initial shape.}
\item{\code{"AtoB"}}{Simulation converges towards the target shape(s).}
\item{\code{"AtoMult"}}{Simulation allows variation but remains close to multiple targets (e.g., full morphospace).}
\item{\code{"Free"}}{Simulation is unconstrained and may deviate freely from known shapes.}
}}

\item{sim}{Integer. Number of simulation iterations.}

\item{npts}{Integer. Number of points to reconstruct each shape.}

\item{a}{Numeric. Phase local threshold to smooth changes. Default is 0.2.}

\item{e}{Numeric. Procrustes distance threshold for acceptance. Default is 0.15.}

\item{dynamic_e}{Logical or data frame. If \code{TRUE}, modifies distance threshold \code{e} dynamically during simulation based on a two-column data frame with time steps and threshold values. Default is \code{FALSE}.}

\item{f}{Numeric. Factor dividing the covariance matrix.}

\item{int.allowed}{Logical. Whether to allow line intersections in candidate shapes. Default is \code{FALSE}.}

\item{only.shapes}{Logical. If \code{TRUE}, returns only simulated shapes at each time step. If \code{FALSE}, returns detailed info including parameters and distances. Default is \code{FALSE}.}

\item{max.attempts}{Integer. Maximum attempts allowed without success in meeting \code{e} threshold before simulation stops. Default is 500.}
}
\value{
A list containing simulated shapes (and optionally parameters and distance metrics) for each iteration.
}
\description{
Simulates shapes over \code{sim} iterations based on a covariance matrix derived from a morphospace.
The simulation starts from an initial shape and optionally converges or varies relative to target shapes.

The morphospace includes amplitude and phase values of shapes. Users can add shapes not originally in the morphospace by providing their amplitude and phase values.
Note that added shapes are excluded from the covariance matrix unless explicitly included before covariance calculation.
}
