% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_morphospace.R
\name{valid_morphospace}
\alias{valid_morphospace}
\title{Check matrix validity}
\usage{
valid_morphospace(x, k, rnd, plotvalues = T)
}
\arguments{
\item{x}{The matrix with the full morphospace}

\item{k}{the number of eigen values to include. Must be higher than 1}

\item{rnd}{Number of randomisations. Default is 100}

\item{plotvalues}{Boolean. Whether the user wants to visualise also the eigenvalues. If F, it only plots the dot product of the eigenvectors and the minimum eigenvalues. Default is T}
}
\value{
A plot showing the directions of the eigenvectors and the minimum eigenvalues. If \code{plotvalues = TRUE} it also shows the k eigenvalues
}
\description{
Given a full covariance matrix, it assesses how many samples are necessary, so that it is positive definite
}
